Fix Betty error

+++ 0x0B-malloc_free/101-strtow.c
@@ -6,13 +6,13 @@
  * strtow -split string in to two
  *
  * @str: Takes string
- * 
+ *
  * Return: pointer or NULL if str is NULL or NULL if fail
  */
 char **strtow(char *str)
 {
 	char **split;
-	long unsigned int i, k;
+	int i, k;
 	int count = 0;
 
 	i = k = 0;
@@ -35,12 +35,12 @@ char **strtow(char *str)
 		if ((str[i] == ' ') || (str[i] == '\t'))
 		{
 			split[i][k] = '\n';
-			k++;	
+			k++;
 		}
 		split[i][k] = str[i];
 	}
-	split[i][k] = '\0'; 
+	split[i][k] = '\0';
 
 	return (split);
-	
+
 }
